+++ src/pages/home/Music.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Music = (props) => (
-    <div {...props} id="musica" className="h-screen bg-isabelline">
-        Musica
-    </div>
-);
-
-export default Music;